backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
black==24.8.0
certifi==2024.12.14
click==8.1.7
dj-database-url==2.2.0
Django==5.1.4
django-cors-headers==4.4.0
django-filter==24.3
django-simple-history==3.7.0
django-stubs-ext==5.1.2
django-types==0.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-redesign==0.4.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
googletrans==4.0.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.0.1
types-psycopg2==2.9.21.20250121
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
